requirements.txt
netifaces
paho-mqtt
flask
sqlalchemy